+++ crates/nix-utils/include/nix.h
@@ -32,8 +32,7 @@ void set_verbosity(int32_t level);
 bool is_valid_path(const StoreWrapper &wrapper, rust::Str path);
 InternalPathInfo query_path_info(const StoreWrapper &wrapper, rust::Str path);
 void clear_path_info_cache(const StoreWrapper &wrapper);
-long unsigned int compute_closure_size(const StoreWrapper &wrapper,
-                                       rust::Str path);
+uint64_t compute_closure_size(const StoreWrapper &wrapper, rust::Str path);
 rust::Vec<rust::String> compute_fs_closure(const StoreWrapper &wrapper,
                                            rust::Str path, bool flip_direction,
                                            bool include_outputs,
@@ -50,17 +49,16 @@ void copy_paths(const StoreWrapper &src_store, const StoreWrapper &dst_store,
                 rust::Slice<const rust::Str> paths, bool repair,
                 bool check_sigs, bool substitute);
 
-void import_paths(const StoreWrapper &wrapper, bool check_sigs, size_t runtime,
-                  size_t reader,
-                  rust::Fn<size_t(rust::Slice<uint8_t>, long unsigned int,
-                                  long unsigned int, long unsigned int)>
-                      callback,
-                  size_t user_data);
-void import_paths_with_fd(const StoreWrapper &wrapper, bool check_sigs, int fd);
-void export_paths(
-    const StoreWrapper &src_store, rust::Slice<const rust::Str> paths,
-    rust::Fn<bool(rust::Slice<const uint8_t>, long unsigned int)> callback,
-    size_t userdata);
+void import_paths(
+    const StoreWrapper &wrapper, bool check_sigs, size_t runtime, size_t reader,
+    rust::Fn<size_t(rust::Slice<uint8_t>, size_t, size_t, size_t)> callback,
+    size_t user_data);
+void import_paths_with_fd(const StoreWrapper &wrapper, bool check_sigs,
+                          int32_t fd);
+void export_paths(const StoreWrapper &src_store,
+                  rust::Slice<const rust::Str> paths,
+                  rust::Fn<bool(rust::Slice<const uint8_t>, size_t)> callback,
+                  size_t userdata);
 
 rust::String try_resolve_drv(const StoreWrapper &wrapper, rust::Str path);
 } // namespace nix_utils
